+++ docs/changelog.md
@@ -21,6 +21,10 @@ Added:
 
 - Add config option `reliable_elevation_measurements` (defaulting to true) that allows hiding the elevation gain data if the user knows that their measurement of elevation data is unreliable.
 
+Fixed:
+
+- Saving privacy zones now works again. ([GH-287](https://github.com/martin-ueding/geo-activity-playground/issues/287))
+
 ## Version 0.44.0 — 2025-05-24
 
 Added: